// src/lib/enums.ts
export const TenantType = {
  INDIVIDUAL: 'INDIVIDUAL',
  COMPANY: 'COMPANY'
} as const;

export type TenantType = typeof TenantType[keyof typeof TenantType];

export const GuarantorType = {
  NONE: 'NONE',
  JOINT_OBLIGOR: 'JOINT_OBLIGOR',
  AVAL: 'AVAL',
  BOTH: 'BOTH'
} as const;

export type GuarantorType = typeof GuarantorType[keyof typeof GuarantorType];